+++ member/new_msg.php
@@ -23,27 +23,8 @@ $info = mysqli_fetch_array($info_res);
 <meta name="description" content="<?=$info['description']?>" />
 <title><?=$info['site_name']?></title>
 <link rel="stylesheet" type="text/css" href="../css/shop_layout.css" />
-<!-- category -->
-<link rel="stylesheet" type="text/css" href="../css/ddsmoothmenu.css" />
-<!--[if lte IE 7]>
-<style type="text/css">
-html .ddsmoothmenu{height: 1%;} /*Holly Hack for IE7 and below*/
-</style>
-<![endif]-->
-
-<?php include_once "../include/inc_script.html"; ?>
-
 <script language="JavaScript" src="../js/global.js"></script>
 <script language="JavaScript" src="../js/member.js"></script>
-<!-- <script type="text/javascript" src="../js/jquery-1.2.6.pack.js"></script> -->
-<script type="text/javascript" src="../js/ddsmoothmenu.js">
-/***********************************************
-* Smooth Navigational Menu- (c) Dynamic Drive DHTML code library (www.dynamicdrive.com)
-* This notice MUST stay intact for legal use
-* Visit Dynamic Drive at http://www.dynamicdrive.com/ for full source code
-***********************************************/
-</script>
-<!-- category end -->
 <script language="JavaScript" type="text/JavaScript">
 <!--